fine funzionalita admin, manca sono aggiornamneto database

+++ src/app/pages/welcome-admin/welcome-admin.component.ts
@@ -19,18 +19,34 @@ export class WelcomeAdminComponent {
 
   ngOnInit(){
     this.authSvc.getAllUtenti().subscribe((users) =>{
-      this.iUsers = users.response
+      this.iUsers = users.response.filter(user => user.ruolo !== "ADMIN");
     })
   }
 
   deleteUtente(id:number){
-    this.authSvc.deleteUtente(id).subscribe(()=>{
-      this.iUsers = this.iUsers.filter(user => user.id != id)
-      this.swal.fire({
-        title: "Good job!",
-        text: "Utente eliminato con  successo!",
-        icon: "success"
-      })
+
+    this.swal.fire({
+      title: "Sei sicuro?",
+      text: "Premi su Cancel per tornare indietro!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#d33",
+      cancelButtonColor: "#C65535",
+      cancelButtonText:"Chiudi",
+      confirmButtonText: "Si, voglio cancellarlo!"
+    }).then((result: { isConfirmed: boolean }) => {
+      if (result.isConfirmed) {
+        this.authSvc.deleteUtente(id).subscribe(()=>{
+          this.iUsers = this.iUsers.filter(user => user.id != id)
+          this.swal.fire({
+          title: "Cancellato!",
+          text: "Utente cancellato corettamente.",
+          icon: "success"
+        });
+        })
+      }
     })
+
+
   }
 }
